requirements.txt
fastapi==0.115.4 
uvicorn[standard]==0.35.0 
sqlalchemy==2.0.36  
pydantic==2.11.7 
pydantic-settings==2.10.1  
python-dotenv==1.0.1  
email-validator==2.2.0  
psycopg2
pyjwt==2.8.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
bcrypt==4.0.1
aiosmtplib==1.1.6